Refactor settings helpers into separate modules
